// client/src/pages/Dashboard/CompetitorDashboard/components/MentalHealtForm.styles.ts
import styled from "@emotion/styled";

export const FormWrapper = styled.div`
  background-color: #121212;
  max-width: 500px;
  width: 100%;
  text-align: center;
  margin: 0 auto;
  padding: 2rem;
  border-radius: 8px;

  @media (max-width: 768px) {
    max-width: 90%;
    padding: 1.5rem;
  }
`;

export const Heading = styled.h1`
  font-size: 1.8rem;
  color: #ffffff;
  margin-bottom: 1.5rem;
  font-weight: 600;
`;

export const Form = styled.form`
  display: flex;
  flex-direction: column;
  gap: 1rem;
  text-align: left;
`;

export const MoodSection = styled.div`
  margin-bottom: 2rem;
`;

export const ZoneGroup = styled.div`
  display: grid;
  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
  gap: 1rem;
  //   margin-bottom: 1rem;
  border: 1px solid #444444;
  padding: 1rem;
`;

export const ZoneLabel = styled.label`
  font-size: 0.9rem;
  color: #ffffff;
  margin-bottom: 0.5rem;
  display: block;
`;

export const Input = styled.input`
  flex: 1;
  background: none;
  border: none;
  outline: none;
  color: #ffffff;
  font-family: ${({ theme }) => theme.typography.fontFamily};
  font-size: 1rem;
  padding: 0.75rem 0;
  width: 100%;
  border-bottom: 1px solid #444444;

  &::placeholder {
    color: #666666;
  }

  &:focus {
    border-bottom: 1px solid #6a5acd;
  }

  &:-webkit-autofill {
    background-color: transparent !important;
    -webkit-text-fill-color: #ffffff !important;
    transition: background-color 5000s ease-in-out 0s;
  }

  &::-webkit-inner-spin-button,
  &::-webkit-outer-spin-button {
    -webkit-appearance: none;
    margin: 0;
  }
`;

export const EmojiGroup = styled.div`
  display: flex;
  justify-content: center;
  gap: 1rem;
  margin-top: 1rem;

  button {
    font-size: 2rem;
    border: none;
    background: none;
    cursor: pointer;
    transition: transform 0.2s ease-in-out;

    &:hover {
      transform: scale(1.2);
    }

    &.primary {
      background-color: #6a5acd;
      color: white;
    }

    &.secondary {
      background-color: transparent;
      color: #6a5acd;
      border: 1px solid #6a5acd;
    }
  }
`;

export const ButtonGroup = styled.div`
  display: flex;
  gap: 1rem;
  //   margin-top: 2rem;

  @media (max-width: 768px) {
    flex-direction: column;
  }
`;

export const Button = styled.button<{ variant: "primary" | "secondary" }>`
  flex: 1;
  background: ${({ variant }) =>
    variant === "primary"
      ? "linear-gradient(135deg, #6a5acd, #483d8b)"
      : "transparent"};
  color: #ffffff;
  padding: 14px;
  border: ${({ variant }) =>
    variant === "primary" ? "none" : "1px solid #6a5acd"};
  border-radius: 6px;
  cursor: pointer;
  font-size: 1.1rem;
  font-weight: 600;
  transition: all 0.3s ease;

  &:hover {
    background: ${({ variant }) =>
      variant === "primary"
        ? "linear-gradient(135deg, #483d8b, #6a5acd)"
        : "rgba(106, 90, 205, 0.1)"};
  }
`;

export const TextArea = styled.textarea`
  background: none;
  border: none;
  outline: none;
  color: #ffffff;
  font-family: ${({ theme }) => theme.typography.fontFamily};
  font-size: 1rem;
  padding: 0.75rem;
  width: 90%;
  min-height: 100px;
  resize: vertical;
  border-bottom: 1px solid #444444;

  &::placeholder {
    color: #666666;
  }

  &:focus {
    border-bottom: 1px solid #6a5acd;
  }
`;