// client/src/constants/views.ts
export enum Views {
  CALENDAR = "calendar",
  ASSIGN = "assign",
  PLAN = "plan",
  PROFILE = "profile",
  MY_PROFILE = "my_profile",
  COACH_PROFILE = "coach_profile",
  REQUESTS = "requests",
  COMPETITORS = "competitors",
  COACHES = "coaches",
  HOME = "home",
  TYPE_CATEGORY_CREATE = "type_category_create",
  TRAINING_TYPE_LIST = "training_type_list",
  MENTAL_HEALTH = "MENTAL_HEALTH",
  MENTAL_HEALTH_CREATE = "MENTAL_HEALTH_CREATE",
  MY_MENTAL_HEALTH_HISTORY = "MY_MENTAL_HEALTH_HISTORY",
  MENTAL_HEALTH_HISTORY = "MENTAL_HEALTH_HISTORY",
  TRAINING_EDIT = "TRAINING_EDIT",
  TRAINING_DETAIL = "TRAINING_DETAIL",
  TRAINING_RESULTS = "TRAINING_RESULTS",
  TRAINING_RESULTS_EDIT = "TRAINING_RESULTS_EDIT",
  CLUB_LIST = "CLUB_LIST",
  CLUB_CREATE = "CLUB_CREATE",
  BOAT_CREATE = "BOAT_CREATE",
  BOAT_LIST = "BOAT_LIST",
  DISCIPLINE_CREATE = "DISCIPLINE_CREATE",
  DISCIPLINE_LIST = "DISCIPLINE_LIST",
}